Enable nested searching
Previous queries are added to the search forms on the results page so
that history is preserved and in theory searches can be nested
infinitely.

To simplify the advanced search logic, in this change the advanced
search form no longer renders search terms into viewing boxes; this
functionality will be added back using Ruby in a future commit.