Show searches by IP address - fixes #19
Without authentication this is the least invasive way to show searches
per user, across different browsers and devices. It means that all
devices in an institution will be able to see each others searches, but
it is more anonymous than cookies. A cookie-based solution would also
show searches per-device, which is not useful for someone who uses
(for example) a phone, tablet and laptop.